# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pathfinder API",
    "description": "AI-powered personalised learning path recommender",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/profile": {
      "post": {
        "summary": "Update Profile",
        "description": "Upsert display name in the profiles table on login/signup.",
        "operationId": "update_profile_api_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events": {
      "post": {
        "summary": "Record Event",
        "description": "Ingest a telemetry event from frontend interactions (quizzes, reviews, completion).\nAccrues mastery in persistent state when applicable.",
        "operationId": "record_event_api_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Record Event Api Events Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/diagnostic/generate": {
      "post": {
        "summary": "Generate Diagnostic",
        "description": "Generate adaptive diagnostic questions targeting highest probe-priority skills\n(skills with high uncertainty and high downstream prerequisite fan-out).",
        "operationId": "generate_diagnostic_api_diagnostic_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnostic/submit": {
      "post": {
        "summary": "Submit Diagnostic",
        "description": "Evaluate diagnostic responses, compute 2PL-IRT theta, and update learner state.",
        "operationId": "submit_diagnostic_api_diagnostic_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagnosticAnswerItem": {
        "properties": {
          "skill_id": {
            "type": "string",
            "title": "Skill Id"
          },
          "discrimination": {
            "type": "number",
            "title": "Discrimination",
            "default": 1.0
          },
          "difficulty": {
            "type": "number",
            "title": "Difficulty",
            "default": 0.0
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "is_correct"
        ],
        "title": "DiagnosticAnswerItem"
      },
      "DiagnosticGenerateRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "title": "Goal",
            "default": "Machine Learning Engineer"
          },
          "num_questions": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 2.0,
            "title": "Num Questions",
            "default": 4
          }
        },
        "type": "object",
        "title": "DiagnosticGenerateRequest"
      },
      "DiagnosticGenerateResponse": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticQuestion"
            },
            "type": "array",
            "title": "Questions"
          },
          "skills_tested": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Tested"
          }
        },
        "type": "object",
        "required": [
          "questions",
          "skills_tested"
        ],
        "title": "DiagnosticGenerateResponse"
      },
      "DiagnosticQuestion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "skill_id": {
            "type": "string",
            "title": "Skill Id"
          },
          "skill_name": {
            "type": "string",
            "title": "Skill Name"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "correct_index": {
            "type": "integer",
            "title": "Correct Index"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "discrimination": {
            "type": "number",
            "title": "Discrimination",
            "default": 1.0
          },
          "difficulty": {
            "type": "number",
            "title": "Difficulty",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "id",
          "skill_id",
          "skill_name",
          "question",
          "options",
          "correct_index",
          "explanation"
        ],
        "title": "DiagnosticQuestion"
      },
      "DiagnosticSubmitRequest": {
        "properties": {
          "responses": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticAnswerItem"
            },
            "type": "array",
            "title": "Responses"
          },
          "goal": {
            "type": "string",
            "title": "Goal",
            "default": "Machine Learning Engineer"
          }
        },
        "type": "object",
        "required": [
          "responses"
        ],
        "title": "DiagnosticSubmitRequest"
      },
      "DiagnosticSubmitResponse": {
        "properties": {
          "theta": {
            "type": "number",
            "title": "Theta"
          },
          "standard_error": {
            "type": "number",
            "title": "Standard Error"
          },
          "updated_mastery": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Updated Mastery"
          },
          "readiness_pct": {
            "type": "integer",
            "title": "Readiness Pct"
          }
        },
        "type": "object",
        "required": [
          "theta",
          "standard_error",
          "updated_mastery",
          "readiness_pct"
        ],
        "title": "DiagnosticSubmitResponse"
      },
      "EventIn": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/EventType"
          },
          "skill_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Id"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "minutes_spent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes Spent"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "EventIn"
      },
      "EventType": {
        "type": "string",
        "enum": [
          "resource_started",
          "resource_completed",
          "quiz_attempted",
          "review_completed",
          "feedback_given",
          "path_generated",
          "detour_inserted"
        ],
        "title": "EventType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "is_guest": {
            "type": "boolean",
            "title": "Is Guest"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "display_name",
          "is_guest"
        ],
        "title": "ProfileResponse"
      },
      "ProfileUpdateRequest": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "ProfileUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}